Source/requirements.txt
# Flask & API
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7

# Machine Learning & NLP
scikit-learn==1.3.0
nltk==3.8.1
pandas==2.1.1
numpy==1.26.0

# Deep Learning (optional, for advanced models later)
tensorflow==2.20.0
torch==2.8.0
transformers==4.56.1

# Testing
pytest==8.4.2
